Change URL to use /calspec/ instead of /current_calspec/
This is because the /calspec contains all past and present data
whereas the current only contains the latest version, so provides
a moving target. If we use the current version, not only do the
versions we use end up changing, but if a star gets updated and we
have not updated the csv file, we will get a download error.